Replace Chunk.len() with Chunk._len
Odd API design choice: public, not-final, read-only, NO-ACCESSOR.

NO-ACCESSOR: This is a high-performance field, and must have a known
zero-cost cost-model; accessors hide that cost model, and make it
not-obvious that a loop will be properly optimized or not.

not-final: set in various deserializers.

Proper usage: read the field, probably in a hot loop.

More javadoc on Chunk
DO TODO: move reportBrokenEnum to Chunk