Fix bug in check_header.  Reading in a file of column headers and all NAs using check_headers=1 would show the correct result in the ParseResult preview, but not affect the final parse.  Some parts of final parse still relied on a parse setup guess.  All non-numeric datasets can't correctly guess if a header exists, so this file exposed the bug.
Final "local" guess of setup has been removed.  Also a bug affecting the preview of data in Flow has been fixed.  The last line of the preview was never being shown.

testXLSBadArgs has been altered to check that a correct exception is thrown.  The parse should not return a 0 row, 0 col result if parse fails.