Basic user-mode swap test
Force odd-ball like-low-memory situations and verify the cleaner writes
to disk, and can read the Chunks back in later.
Doesn't actually trigger a GC (so does not test, e.g. the GC heuristics
for caching levels).
DOES check that ice_root can be written, read, deleted, and is written
and read.