Print a handy message (Open H2O Flow in your web browser) when the cluster comes up like Sparkling Water does.
To prevent lots of messages crossing each other, don't print this until all
the nodes have checked in with the full cluster size.  The code used to
"finish" after the first node checked in with the full cluster size.
This should not make any difference to the user, both behaviors are correct.