Fixed orderly shutdown to work with flatfile.
2 fixes:

a) udp packet handling was skipping basic_packet_handling method in UDPReceiver thread which did the check for suicide message.

b) Turned shutdown into its own task, which is send from the driver node to all other nodes, where it starts a timeout which will force nod determination. The driver node then tracks which nodes ack the task (confirm that they are switching off), the nodes will proceed with shutdown when they either receive ackack for the shutdown task or the timeout expires.