Fix download of pojo in Python.
On Windows the target file for h2o-genmodel.jar has to be opened with 'b' flag.

From Python doc:

"On Windows, 'b' appended to the mode opens the file in binary mode, so
there are also modes like 'rb', 'wb', and 'r+b'. Python on Windows makes
a distinction between text and binary files; the end-of-line characters
in text files are automatically altered slightly when data is read or
written. This behind-the-scenes modification to file data is fine for
ASCII text files, but it’ll corrupt binary data like that in JPEG or EXE
files. Be very careful to use binary mode when reading and writing such
files. On Unix, it doesn’t hurt to append a 'b' to the mode, so you can
use it platform-independently for all binary files."