Make a special Null for Vecs
Vecs have a "deleted value" similar to a TOMBSTONE in NBHM.
This value supports a count of outstanding invalidates, and is used to
prevent a DKV remove, followed quickly by a racey DKV.put to install in
the DKV while the delete is still outstanding.
Change the rwlock invariant to start with a count of 1 for both local
and remote Values.  Home inserts need the count of 1, to prevent a racey
2nd home insert from instantly removing the Value before we can track
any pending invalidates.