Set codename and version number for tibshirani release
Codename: tibshirani
Version: 3.6.0.X

Cite https://en.wikipedia.org/wiki/Robert_Tibshirani:

"Robert Tibshirani (born July 10, 1956) is a Professor in the
Departments of Statistics and Health Research and Policy at Stanford
University. He was a Professor at the University of Toronto from 1985 to
1998. In his work, he develops statistical tools for the analysis of
complex datasets, most recently in genomics and proteomics.

His most well-known contributions are the LASSO method, which proposed
the use of L1 penalization in regression and related problems, and
Significance Analysis of Microarrays. He has also co-authored three
well-known books: "Generalized Additive Models", "An Introduction to the
Bootstrap", and "The Elements of Statistical Learning",[1] the last of
which is available for free from the author's website."