Fixed deadlock in RollupStats computation.
RollupStats can be requested multiple times in parallel, RollupStat's home node will launch only one task for the first thread, all others will block for the task to be finished.
The task to compute rollupstats was being launched in non-blocking (not-consuming the first thread) way using h2o counted completer and keeping the same priority level, however, all the subsequeqnt request to rollupstats would block a thread - most likely on the same priority level. This lead to the following sequence of events and deadlock:

1) first thread comes in, acquires lock (via DKV) and makes task to compute rollup stats, but the task is not running yet, it is submitted to global queue on the same priority level.
2) n other threads come requesting the same rollupstats computation on the same priority queue, find out that the task for it already exists and block for it, blocking their threat and depleting the thread pool at the given priority level
3) there are no threads to compute the rollup stats -> deadlock.

The fix is to raise the priority of the task computing the rollupstats by 1. Switched to blocking call, no need to use completer interface (although possible, not worth it).
Because we're raising the priority by 1 (and MRTasks can not be launch on priority queues with priority >= MIN_HI_PRIORITY), had to also decrease priority of user-invoked rollups by 1.