PUBDEV-2776: Implement the ability to set the weights/biases for each layer of a DL model.
This can be used to make your own stacked autoencoders, port over models from
other tools (and get a POJO), and to play with your own weight generation
tools.  You can train a model with epochs=0, to make sure that the weights
won't be altered, if you want to just see how that models performs.