[PUBDEV-2897, PUBDEV-2893] Fix a bug in the parser.
The parser's domain collector (the task `CreateParse2GlobalCategoricalMaps`) and the following task `UpdateCategoricalChunksTask`
did not share the same strategy to identify categorical columns.

Hence, in some scenarios demonstrated by dataset `smalldata/jira/pubdev_2897.csv`
it resulted into a situation when the `CreateParse2GlobalCategoricalMaps` task missed
a categorical column (because of shared categorical domain was empty) and did not compute mapping, but also computed
wrong mapping index for the following categorical column.

Hence, the following task obtained
a misaligned array of mappings. At the end, the mapping was used but for a different column resulting
into wrong values inside vector.