bugfix with sparse categoricals (#53)
* Bugfix with sparse categoricals

1st bug: after flipping to sparse after adding 16 0 cat's, the 17th (and
later) 0 cat was not added sparsely.
2nd bug: in a column with mixed sparse 0's and non-sparse 0's, the cat
type was lost even if only ever "NewChunk.addCategorical") was called,
and the chunk was declared numeric... and the non-sparse 0 cat's got
converted to NA's.  Later the whole column could still be declared
categorical.  Effect was to convert a bunch of 0's to NA's.

* missed file