Updated Value to not to cache raw bytes. Raw bytes are stored only if there is no mojo and vice versa.
This is in preparation for storing chunks in blocks.
Right now chunks are supported by the raw bytes. If we store chunks in blocks, the natural solution is to simply store array of chunks in each block.
However, this will loose the property of being supported by raw bytes and thus we would double the data size.

Since the raw bytes do not have to be cached at present, there is no change in semantics, the only possible drawback is in degraded performance.
I do not expect performance to be affected since each key will be mostly send to N nodes (nodes cache the remote keys). Still will measure before pushing back to master.