[SW-57] Produce artifacts for Scala 2.11 and Scala 2.10 (#70)
The build system was changed to produce artifacts for two major
versions of Scala. The implementation is based on a Gradle plugin for
which is applied to `settings.gradle` and for each project which
need cross compilation it introduces two new projects suffixed by base Scala
version. For example: for `h2o-scala` it produces `h2o-scala_2.10` and `h2o-scala_2.11`.

Additionally, each scala project can have different Scala sources - a shared one is `src/main/scala`
and a specific one in `src/main/scala_2.10`.

The commit also:
  - cleans up generated POM file with updated information.
  - remove unused classes from Scala API.