[WIP] Improves performance of importing many small files into H2O
This is a preview pull request demonstrating the concept.

Performance of H2O suffers considerably when we try to import hundreds
or thousands of input "part" files into a single Frame. This is caused
by H2O's chunk distribution algorithm. Because chunk location is
determined based on chunk index chunks read from small files are not
uniformly spread accross the cluster. They are concetrated
on a single node (or few nodes, depending on the size of part file).
The parsing is then also not distributed but performed by a single node.

The idea of this change is to create a single Vec for all part files.
First chunk of each part file (also every other chunk) has a different
index in the Vec. This leads to better distribution of chunks
and true parallel parsing. From the point of view of import functionality
and also chunk distribution there is no difference between working
with bunch of files and a single large continuos file.

This concept is limited for HDFS files as of now. In can be generalized
and pushed down to FileVec level.