Disable parallel parse if input is large amount of small files
Parsing large amount of small files is inefficient in H2O. This is caused
by the chunk distribution function. Chunks with the same indices are always
stored on the same node. If we have many small files, we will end-up with
most of the chunks on first few nodes of the cluster (regardles of the size
of the cluster). This change tries to even out the amount of work each node
of the cluster does by disabling parallel parse of small files and relying
on distributing the work more uniformly using the keys of the files.