PUBDEV-3505 Fixed handling of (na) sparse uuid columns, added test.
UUIDs currently do not support sparse. If the column starts with bunch of NAs it is stored as na-sparse numbers initially.
It is flipped to uuids after hitting first non-na uuid value, however, at this point, the parse was not cancelled.
This lead to unsupported state where part of data was stored sparse while the rest was dense -> lead to AIOB.

Fixed by cancelling the sparse when flipping to UUIDs.