Code serialization works in the case of upper-level class.
Added a function-based generator to Vec, so we can pass a function and generate data.

TODO: try to make it work with inner classes; wrap it neatly, so that Closure is transparent. Then implement functional vectors (we are almost there). Maybe add other classes that implement Function (talk with Michał).