PUBDEV-3610: Remove random-split-points
Removed random-split-points option from GBM and DRF sections of the
user guide. Also removed this option from the GBM booklet.
Driveby fixes: Fixed typo in fold_assignment description. Alphabetized
current list of parameters in the appendix. Created links to the
current parameters in the appendix from the GBM section of the user
guide.