Improves parsing capabilities of ModelMojoReader
Existing implementation of readkv(String) has to guess how should be
a string value parsed. This is not always desirable because a missing value
or incorrect guess can lead to a NPE or ClassCastException.

The parser had trouble with recognizing a string value "[]" as a list of
integers. To guide the parser we provide a default value which is used
to derive the correct result type.

This is a quick solution. More work needs to be done to avoid unexpected
behavior.