Documentation: Added quasi_binomial family option in GLM (#598)
* Documentation: Added quasi_binomial family option in GLM

PUBDEV-3482 adds “quasi_binomial” family option in GLM. Added this to
the GLM section of the UG and to the GLM booklet.
Driveby fixes: Added several parameters that were missing from the GLM
booklet. Also added “multinomial” to list of link options.

* minor update

Added “Multinomial” to list of available link options.

* Renamed new option

New option is now “quasibinomial”. Also removed “multinomial” as an
option for link. This is specified using family_default.