PUBDEV-3703
Added detailed descriptions and examples of the remaining GLM
parameters for the Appendix section.
Driveby fixes: Fixed typos in glm.rst file and the in “epsilon” and
nlambdas parameters. Added a note in the lambda_search parameter that
GLM will calculate the minimum lambda value unless a lambda is
specified. Fixed alphabetized list for the distribution parameter.