PUBDEV-3694: PCA with wide dataset when pca method is set to be GramSVD. (#672)
First, implemented capability for PCA to deal with wide datasets with many columns but fewer rows.  I used the trick
of finding the transpose of the PCA.  Next, we multiple the eigenvectors with the correct data matrix to obtain the
PCA of the wide dataset.

Second, I implemented Junit tests (PCATest.java) to compare the PCA obtained on a small dataset using the original and
my wide PCA method to make sure they are the same.

Third, added Erin's Rotterdam test as a Runit and Pyunit test.  I compare our wide PCA results with R result and make sure
they are close.

Thanks to Tomas Nykodym code review suggestions.  I found 3 bugs out of it.

PCA with wide data set.  Put in Tomas fix to solve errors when running in multi-node mode.