Moved out all "string to bytes" transformatios
Moved them into a `StringUtils` method.
The thing is, repeating "udf, udf" everywhere looks kind of too repetitive.
Using default encoding means we depend on whatever settings a server has,
and the server can be something in Japan (and another in Russia, and yet another in China), so we can't seriously rely on everything being UTF-8.

So there.

A method `toBytes()` is added to apply to whatever object comes up.
Tests passed on my machine.