PUBDEV-4315
Even though zero columns should be filtered up front before the algorithm is even run, they can happen due to e.g. categorical level never occurring in the given fold. When this happens, coefficient should be removed on the fly and set to 0. However, this does not quite work for multinomial - there is too much complexity in working with subset of coefficients at a given time. Since multinomial is currently only implemented via coordinate descent over classes, it should be run with regularization and so zeroed predictors are not a problem.

Fix: 1) do not remove zeroed predictors if family==multinomial.
       2) refined zeroed column detection - require both XtX and XtY to be zeroes for the given column. Zero XtX can happen in case of e.g. running binomial and given predictor turning prediction to 1 whenever non-zero. Zero XtY trivial occurs if response is zero for every non-zero predictor value. In both cases, predictor should be left in.