PUBDEV-4549: Add supported aggregate functions in R (#1260)
* PUBDEV-4549: Add supported aggregate functions in R

Added the supported aggregate functions for the group_by function. This
information is in the Details section of the group_by function.

* Added param na.method to group_by function

* Simplified rm.method parameter description

Most of this information was already included in the Details section.

* Added nrow parameter

* Changed rm.method to na.method in list of arguments.

* Edits after receiving more info

Removed nrow and na.method from the list of arguments and included
those in the Details section only.