PUBDEV-4583: Updated description of alpha param (#1261)
* PUBDEV-4583: Updated description of alpha param

In glm.R, updated the description of the alpha parameter.

* Updated the GLMV3.java file

The change to the alpha description has to be made in the GLMV3.java
file. Built H2O locally after, and the associated files were also
updated. Also, the generated PDF displays the updated description for
alpha.
Driveby fix: Fixed capitalization in other GLM parameter descriptions.