PUBDEV-4583: Updated description for Alpha command in R docs
Updated the GLMV3.java file with an improved description for the alpha
command.
The change to the alpha description has to be made in the GLMV3.java
file. Built H2O locally after, and the associated files were also
updated. Also, the generated PDF displays the updated description for
alpha.
Driveby fix: Fixed capitalization in other GLM parameter descriptions.