PUBDEV 4213 rel vajda - Updates to markdown file syntax (#1303)
* PUBDEV-4212 - Fixing markdown files - WIP

Fixing formatting in old markdown files. WIP.

* PUBDEV-4213: Committing updates from the "howto" folder

updated the markdown files in the “howto” folder.

* Updates after checking in

Additional updates in the “howto” folder after checking in and testing.

* updates to tutorials folder

Updated the markdown formatting for md files in the tutorials folder.
Driveby fix: Fixed typo in the Interpreting the Gains/Lift Chart in the
User Guide.

* changes after review

One update after reviewing checked-in markdown files.

* Changes to markdown files in the "upgrade" folder

Updated the markdown formatting for md files in the “upgrade” folder.

* Changes after review

Minor updates after reviewing checked-in markdown files.

* Updates in the Flow folder

Updated the markdown files in the “flow” folder.