PUBDEV-4733: Document default values for XGBoost parameters (#1413)
* PUBDEV-4733: Document default values for XGBoost parameters

- Updated the XGBoost parameters to include default values.
- Compared H2O default values with XGBoost defaults. Created
PUBDEV-4751, PUBDEV-4752, and PUBDEV-4573, which describe
inconsistencies.
- Added aliases for options that can be used interachangeably. However,
the aliases still have different default values. I will update those
default values after the defaults are fixed in the code.
Driveby fixes: Updated the default values for stopping_rounds and
stopping_tolerance in the Parameters appendix.

* PUBDEV-4733: Driveby fix for XGBoost MOJO and POJO info in Flow

MOJOs are supported for XGBoost, and this info was missing in the Flow
topic. Updated list of MOJO-supported algorithms in the Flow topic and
indicated that POJOs are not supported for XGBoost.