PUBDEV-4889: Updated description of weights_column (#1543)
* PUBDEV-4889: Updated description of weights_column

Added more text to the description of the weights_column parameter to
match what exists in the user guide. This is in response to support
tickets 91136. I think that ModelParaemterSchemaV3.java was the correct
place to make the change.

* additional updates

Built H2O to generate updated files. This should prevent errors in the
startup_checks test.