PUBDEV-5056: Binary models not compatible across h2o versions (#1718)
Added a note to the “Saving and Loading a Model” and “Faq > General”
sections describing that models saved in Python, R, or Flow using one
version of h2o cannot be loaded in another version. A workaround is to
use POJO/MOJO.
Driveby fix: Noted in the “Getting Data into H2O” section that this
refers to getting data into your h2o cluster (as opposed to an H2O.ai
cloud).