PUBDEV-4895: Documentation for early stopping (#1738)
* PUBDEV-4895: Documentation for early stopping

Added a section in the Algorithms > Common topic (common.rst)
describing how users can specify early stopping for models built with
supervised learning algorithms (which varies based on the algorithm).
Driveby fix: Set up anchors and use a reference tag to link to the
relevant Parameters sections. This is better than using HTML links
because they are usable in both the rst file (if a user is in github)
and in the html output.

* Stopping parameters for grid search

the same parameters used by automl, gbm, drf, etc also apply to grid
search. added this to the heading and to the introductory paragraph.