PUBDEV-5429: Add support for ingesting from multiple directories + speedup (#2653)
This change also improves performance of ingest when importing from
large number of source files / directories. Previous implementation
was sequential and resulted in a single request per source. This is
especially bad for HTTP ingest which is done eagerly.

Current implementation does a single request for all the files. Loading
is paralellized on a single node (not distributed - typically not necessary).