PUBDEV-6954: Fix export checkpoint when CV is enabled
The issue is caused by a race condition in the method that tests whether
the target directory for checkpoints is writeable or not.

This can only happen in CV scenario, when models are built in parallel
As a quick fix we are only enabling the check on the main model, CV models
don't need to do the same validation.