PUBDEV-6460: We can use NM_HOST to find out IP of H2O node
This won't be directly exposed via a parameter, instead users can add
an option to H2O driver like this

hadoop jar h2odriver.jar -Dh2o.ip.envvar=NM_HOST -mapperXmx 4g -n 10

This will make H2O lookup the hostname (of the associated node manager)
and use this IP instead of the built-in heuristics.

If H2O is running containerized in Docker, this approach might fail but
we might be able to get the IP from a different property (this is why we
are leaving the setting up to the user).