Update models.R
For partials of 2 variables, h2o.partialPlot is getting hung up when the pp tables' structure changes to more than one variable for partial dependence plotting. The original indexes are set so that the second column is the mean response value but this is not the case in a partial on more than 1 variable. To update this, we'll take the column dimension and index backwards from there. 

In the future this section can probably just be removed because I don't think it directly affects any of the functions that follow.