Publich only Python 3.6 and later to Conda
There is an issue with certifi package that cannot be installed for
2.7 and 3.5 using the conda-forge channel.

A simple fix would be to enable another channel that has the package for
earlier version. Instead of doing that, we decided we will no longer
deploy the older python versions to Conda - they don't seem to be used
and are thus just taking space.