PUBDEV-7138 - use keys reference to CompressedIsolationTree in Extended Isolation Forest instead of the CompressedIsolationTree object.
It reduce time and memory consumption to finish the model and fix the problem with exceptional model finish. For example there was a problem with exceptional completion if cluster run out of memory, where insteaf of the exeption raise, computation finish with endless loop and try to free memory for large
with reference to CompressedIsolationTree objects.