PUBDEV: Fix splits NA vs Rest to be correctly represented
Categorical splits NAvsREST are not represented by a bitset in the
CompressedTrees - there is bunch of places in the code where the bitset is
removed. However, the type of split stays categorical equal = 2 or 3.

This means the scoring logic interprets the split as categorical. While
no bitset is actually parsed from the tree, the scoreTree method can
erronerously re-use a previous bitset split instead of creating a fresh new
one.

This will result in mis-classification of the observations in the tree nodes.