Upgrade to Jetty 9.4.44.v20210927 in the minimal assembly
Create h2o-jetty-9-minimal for use in the minimal assembly. This new Jetty
integration doesn't allow to use any authentication, it is targeted
for standalone single node deployments (=laptop).

WebSockets and other extended Jetty features are not supported.

The point of the minimal versions is to be able upgrade to a newer versions
quickly (Jetty authentication modules need to be carefully selected and
tested). Having no support for authentication in the minimal assembly
simplifies moving forward with Jetty versions.