Disable tests on h2o-test-support - there is no actual test anyway
This was done because currently there was only H2ORunnerPlayground class
that was meant for experimenting on dev machines, not to be run on CI.
This class fails on Java 16/17 because of missing "opens" when JUnits
are executed.

If there were tests added to this module, we would have to solve it in
another way.