Dummy replacement for JndiLookup affected by zero-day expoits
One of Log4j's suggested mitigations is to delete JndiLookup class from
affected jars, from https://logging.apache.org/log4j/2.x/security.html#CVE-2021-44228:

> Log4j 2.x mitigation: Implement one of the mitigation techniques below.
> ...
> - Otherwise, remove the JndiLookup class from the classpath: zip -q -d log4j-core-*.jar org/apache/logging/log4j/core/lookup/JndiLookup.class

H2O produces error messages when this class is removed, to avoid such error
messages we suggest to put a dummy version of the JndiLookup class in place
of the original class. This class will printout the build is not affected
when it is being loaded (in our case that is alway on H2O startup).

This is verification to the users that build is safe to use.