We no longer need SecurityUtils to be compatible with Java 7
SecurityUtils were originally meant to be added to classpath of older
Hadoop builds of h2odriver.jar that had a bug that prevented it from
correctly using -internal_secure_connections option.

Older versions of h2odriver were compatible with Java 7 and thus we
preserved the compatibility with Java 7.

We no longer know of anyone using Java 7 in production and we can abandon
the Java 7 compatibility.