Improve PersistHdfs#wrapSeekable - avoid caching data on non-home node
When parsing Parquet files we need to access the metadata part of the file,
node that owns the chunk corresponding to end the metadata part of the file
will typically be a different a node in the cluster. This could cause
excessive caching of chunks.

In order to avoid that we add a flag to VecDataInputStream that will
immediatelly flush the local cache after the chunk is loaded in memory.