ITA-1033: Developers need to mark PR's impact on model productionizing
In order to be able to alert customers about possible issues with their
production models - we need to have a simple filter that will tell us
whether a change fixed issue with predictions and/or model reproducibility.

We will introduce a field “Production Impact” that each engineer will be
required to fill in order to pass PR checks and merge the PR.

This PR encorces the requirement.