PUBDEV-8757: Explicitly start external xgboost cluster before CV [hotfix]
In this hotfix we attempt to start external xgboost cluster at a preditable
time before the CV models start executing. This prevents the CV model
to compete which one will start the cluster and makes the whole flow
more deterministic.

This is a best effort fix to address the common scenarios, it doesn't fully
fix PUBDEV-8757.