Add "Persist S3" test stage - to test user given security credentials
We set environment variables whose names coincides with AWS credentials.
This makes it impossible to test function like set_s3_credentials because
we can never be sure if the correct credentials were picked from the
environment or if they were taken from the user-provided credentials.

In this commit we introduce a new stage that has the credential variables
prefixed with "H2O_" and AWS SDK cannot just pick them up from environment.