PUBDEV-8802: H2O dependencies should not override Jetty implementation
Adding a dependency on h2o should not introduce classes that could clash
with different Jetty versions. We have a custom implementation of Response
class that we currently bundle in h2o-jetty-9 - this is wrong because
any project adding this dependency will see our Response class instead of
the Response class that is coming from their version of Jetty. To address
this we can only bundle the modified class in the h2o assembly jar.