Remove nparallelism from GLM VIF parameters
It is unlikely to be correctly used by users, especially because it is only
used by VIF but it is a GLM parameter.

The heuristic should be good enough to work well for most cases and user
doesn't need to worry about it.

For special cases users can set a system property that is specific for
VIF on a given training frame. Properties of training frame is really what
drives parallelism and thus specifying on this level could be beneficial.