Workaround conda 3.7+ release issue
I tried switching from `conda-forge` channel to `anaconda` channel and was able to make the build work. The difference between those 2 channels is that `anaconda` has currently the older-style version of the `certifi` package (doesn’t have the “noarch” version yet). But that is a workaround that is probably not good for the future.