XGBoost: Allow to train CV models concurrently on the same GPU
Previously we tried to dedicate a GPU exclusively to one model (not just
one CV model, but any model in general).

This meant that if you start one model (e.g. from Python) asynchronously
and then start a second model - the first model will start training
and only after the it is done the second one will start executing. This
is a behavior that I would not expect as an user.

There is no harm in running multiple models on GPU as long there is enough
memory. While it may not be the best strategy in general, in H2O's case
we have quite a bit of overhead in CPU-only scoring. Having 2 CV models
running concurrently - parallelizes the overhead and saves some time 5-10%
in testing.